requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.11
SpeechRecognition==3.10.1
PyAutoGUI==0.9.54
customtkinter==5.2.2
Pillow==10.2.0
pyaudio==0.2.14
